Fixed tests, realized something about short vs. long.
I was keeping a distinction between short and long options for a
misguided reason.

In Windows, switches are "long options" in Unix.  /abc is always
eqivalent to --abc; if you want something like -abc, you would
type /a/b/c.  So, the only time that a single character matters
is on the Unix side.

I was making the distinction, folding one character strings down to
single characters, and the like.  The idea was that you could use
CHARACTERP or STRINGP to tell when something was a long or short
option.  But this is silly, I now think.

Everything becomes strings.  If the user needs it, they can easily
introduce a function that converts strings to characters.  Sure,
strings have overhead and are slower, but just how many command lines
do you intend to process?

So, check everything in, before I start simplifying everything...