More tests done.
Going to pause here and make the change in styles from a
assoc list to a hash table.  Then we can eliminate the :canon
nonsense and just pass a hash around everywhere.

So, canonicalize-styles becomes hash-styles or something like
that.  A few top-level entries for the conversion, and the
with-styles crap goes away, too.