Much testing fixed/restored.
Among the problems, including silly typos, was one not-so-silly
problem.  The ASDF3 cache got corrupted from when I was trying
to recover the git history a while back.  As best as I can tell,
it seems that something about more than one project named "petulant"
being visited at different times reallllly hosed ASDF3's cache of
what lived where.

It's fixed now, I wish I could have debugged it and sent in a
problem report.  As it is, though, I spent FAR too much time trying
to narrow it down.  Was it in Projectile?  Was it in Roswell?  Was
it somehow coming from Emacs?  When the ASDF3 cache goes weird, you
seriously can't immediately tell where the source of the confusion
is.  Oi.

Anyway, fixed now.