Beginning a new context layer.
Will replace styles with it.  Will make the change reach all the way
down to simple parsers.  Basically, in adapting the top layer to call
the bottom layer, I realize the context stuff needs to be supported
all the way down to the lower functions.  This will simplify the top
layer and avoid needless recomputation of certain lists and other data.