Okay, changing the way we handle keywords.
Starting from the bottom, we'll removing the existing handling of
things like :KEY and :NOKEY.  :NOKEY goes away, that was just
stupid.  We'll introduce a function that's only called when we're
returning a value (or setting it in *OPTIONS*) for rewriting options.
The rest of the time, we'll leave everything as a string and rely on
the hashes to compare strings usefully (as we were doing before).

I got carried away, I think, and overcomplicsted it.

Also, still need to resolve the SIMPLE and PARSE naming issues.
It's functional, but it feels wrong.