From the top down.
Changed asd to use serial components.  I know, it looks lazy :-)
but it actually serves a purpose.  For the time being, it implies
an order for visiting all the files.  This, in turn, makes "top down"
changes easier to trace and edit through the system.

Removed some of the Markdown.  We'll use it structurally, but the
inline `foo` and similar just looked too awkward.  As it is now,
the docs are subtle.  They read plainly just fine, but if you throw
the docstrings through Markdown they come out nicely formatted.  This
is the middle ground I was looking for.

Got rid of "loose" dictionaries and tries; instead, just promoting
a more traditional :TEST interface.

Updating tests as we go along.

Ready to pick up changes to PARSE that we were in previously: namely,
resolving that ugly chgname optname optname1 optname2 business once
and for all.