Still fixing up parse.lisp.
Look at the breakage between windows-f2 and windows-f3; in one
context, the option's argument is NIL, while in the other context
is it an empty string.  This used to work right.  Is it related
to the elimination of (advance) that I tried to edit back in?
Probably.  I bet with more tests with subsequent options on the
command line we can turn it up.

Then, all of parse-1 through parse-5 are broken.  Sigh.

But, after that, parse.lisp is all working again!