Just thought of something very different.
Instead of different functions and everything, let's shoot
for more of a spec, or maybe even DSL, approach that defines
everything.

Checking in everything right now.  Will create an experimental
branch after this, because these are going to be sweeping
changes.  I'll play around in ./wwi first to settle on the new
approach.